Initial stab at letting package.json be source of truth in a Projen project.
Why? To enable the usage of third-party tools that do not yet understand projenrc.ts,
and thus cannot automatically create PR's with upgrades.

Assume the NodeProject's "Dependency Upgrades" GitHub Action is not an option
https://projen.io/docs/project-types/node/#dependency-upgrades

The config here will read package.json before synthesis, and populate the Project
with the dependencies that already exist. This will mean a PR which updates
package.json with a version upgrade will be retained by projen.

It also has the downside than an upgrade of Projen may not automatically update
it's own devDependencies.